Improved DependnecyScanScore
- Updated DependencyScanScore to use DependabotScore
  and OwaspDependencyCheckScore.
- Merged UsesOwaspDependencyScan data provider to UsesOwaspDependencyCheck.
- Updated PrettyPrinter.
- Updated tests and test vectors.

Part of #298
