+++ src/main/java/com/sap/sgs/phosphor/fosstars/data/github/UsesOwaspDependencyCheck.java
@@ -2,33 +2,59 @@ package com.sap.sgs.phosphor.fosstars.data.github;
 
 import static com.sap.sgs.phosphor.fosstars.maven.MavenUtils.browse;
 import static com.sap.sgs.phosphor.fosstars.maven.MavenUtils.readModel;
+import static com.sap.sgs.phosphor.fosstars.maven.ModelVisitor.Location.BUILD;
+import static com.sap.sgs.phosphor.fosstars.maven.ModelVisitor.Location.REPORTING;
+import static com.sap.sgs.phosphor.fosstars.model.feature.oss.OssFeatures.OWASP_DEPENDENCY_CHECK_FAIL_CVSS_THRESHOLD;
+import static com.sap.sgs.phosphor.fosstars.model.feature.oss.OssFeatures.OWASP_DEPENDENCY_CHECK_USAGE;
+import static com.sap.sgs.phosphor.fosstars.model.other.Utils.setOf;
+import static com.sap.sgs.phosphor.fosstars.model.value.OwaspDependencyCheckUsage.MANDATORY;
+import static com.sap.sgs.phosphor.fosstars.model.value.OwaspDependencyCheckUsage.NOT_USED;
+import static com.sap.sgs.phosphor.fosstars.model.value.OwaspDependencyCheckUsage.OPTIONAL;
 
 import com.sap.sgs.phosphor.fosstars.maven.AbstractModelVisitor;
+import com.sap.sgs.phosphor.fosstars.maven.ModelVisitor.Location;
 import com.sap.sgs.phosphor.fosstars.model.Feature;
-import com.sap.sgs.phosphor.fosstars.model.Value;
-import com.sap.sgs.phosphor.fosstars.model.feature.BooleanFeature;
+import com.sap.sgs.phosphor.fosstars.model.ValueSet;
+import com.sap.sgs.phosphor.fosstars.model.feature.oss.OssFeatures;
+import com.sap.sgs.phosphor.fosstars.model.value.OwaspDependencyCheckCvssThresholdValue;
+import com.sap.sgs.phosphor.fosstars.model.value.OwaspDependencyCheckUsage;
+import com.sap.sgs.phosphor.fosstars.model.value.OwaspDependencyCheckUsageValue;
+import com.sap.sgs.phosphor.fosstars.model.value.ValueHashSet;
 import com.sap.sgs.phosphor.fosstars.tool.github.GitHubProject;
-import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.InputStreamReader;
+import java.nio.file.Path;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import org.apache.maven.model.Model;
 import org.apache.maven.model.Plugin;
 import org.apache.maven.model.ReportPlugin;
+import org.codehaus.plexus.util.xml.Xpp3Dom;
 
 /**
- * This data provider checks if an open-source project uses OWASP Dependency Check Maven plugin to
- * scan dependencies for known vulnerabilities.
+ * This data provider checks if an open-source project uses OWASP Dependency Check to scan
+ * dependencies for known vulnerabilities. In particular, it tries to fill out the following
+ * features:
+ * <ul>
+ *   <li>{@link OssFeatures#OWASP_DEPENDENCY_CHECK_USAGE}</li>
+ *   <li>{@link OssFeatures#OWASP_DEPENDENCY_CHECK_FAIL_CVSS_THRESHOLD}</li>
+ * </ul>
  */
-public class UsesOwaspDependencyCheck extends CachedSingleFeatureGitHubDataProvider {
+public class UsesOwaspDependencyCheck extends GitHubCachingDataProvider {
 
   /**
-   * A feature which is filled out by the provider.
+   * A list of configurations associated to OWASP Dependency Check plugin which can fail the build.
    */
-  static final Feature<Boolean> USES_OWASP_DEPENDENCY_CHECK
-      = new BooleanFeature("If a project uses OWASP Dependency Check");
+  private static final Map<String, String> FAIL_BUILD_CONFIGURATIONS = new HashMap<>();
+
+  static {
+    FAIL_BUILD_CONFIGURATIONS.put("failBuildOnAnyVulnerability", "boolean");
+    FAIL_BUILD_CONFIGURATIONS.put("failBuildOnCVSS", "number");
+    FAIL_BUILD_CONFIGURATIONS.put("junitFailOnCVSS", "number"); 
+  }
 
   /**
    * Initializes a data provider.
@@ -40,82 +66,250 @@ public class UsesOwaspDependencyCheck extends CachedSingleFeatureGitHubDataProvi
   }
 
   @Override
-  protected Feature supportedFeature() {
-    return USES_OWASP_DEPENDENCY_CHECK;
+  protected Set<Feature> supportedFeatures() {
+    return setOf(
+        OWASP_DEPENDENCY_CHECK_USAGE, 
+        OWASP_DEPENDENCY_CHECK_FAIL_CVSS_THRESHOLD);
   }
 
   @Override
-  protected Value<Boolean> fetchValueFor(GitHubProject project) throws IOException {
+  protected ValueSet fetchValuesFor(GitHubProject project) throws IOException {
     logger.info("Figuring out if the project uses OWASP Dependency Check ...");
     LocalRepository repository = GitHubDataFetcher.localRepositoryFor(project);
-    boolean answer = checkMaven(repository) || checkGradle(repository);
-    return USES_OWASP_DEPENDENCY_CHECK.value(answer);
+
+    ValueSet values = new ValueHashSet();
+
+    checkMaven(repository, values);
+    if (found(values)) {
+      return values;
+    }
+
+    checkGradle(repository, values);
+    if (found(values)) {
+      return values;
+    }
+
+    values.update(OWASP_DEPENDENCY_CHECK_USAGE.value(NOT_USED));
+    values.update(OWASP_DEPENDENCY_CHECK_FAIL_CVSS_THRESHOLD.notSpecifiedValue());
+
+    return values;
   }
 
   /**
-   * Checks if a project uses OWASP Dependency Check Maven plugin.
+   * Checks if a value set contains the features for OWASP Dependency Check.
+   *
+   * @param values The value set to be checked.
+   * @return True if the value set contains the feature, false otherwise.
+   */
+  private static boolean found(ValueSet values) {
+    return values.has(OWASP_DEPENDENCY_CHECK_USAGE)
+        && values.has(OWASP_DEPENDENCY_CHECK_FAIL_CVSS_THRESHOLD);
+  }
+
+  /**
+   * Checks if a project uses OWASP Dependency Check Maven plugin,
+   * and updates a value set if the plugin is found.
    *
    * @param repository The project's repository.
-   * @return True if the project uses the plugin, false otherwise.
+   * @param values A {@link ValueSet} to be updated.
+   * @throws IOException If something goes wrong.
    */
-  private boolean checkMaven(LocalRepository repository) throws IOException {
+  private static void checkMaven(LocalRepository repository, ValueSet values) throws IOException {
     Optional<InputStream> content = repository.read("pom.xml");
 
-    if (!content.isPresent()) {
-      return false;
-    }
+    if (content.isPresent()) {
+      try (InputStream is = content.get()) {
+        Model model = readModel(is);
 
-    try (InputStream is = content.get()) {
-      Model model = readModel(is);
-      return browse(model, withVisitor()).result;
+        Visitor visitor = browse(model, withVisitor());
+
+        OwaspDependencyCheckUsageValue usage = OWASP_DEPENDENCY_CHECK_USAGE.value(visitor.usage());
+
+        OwaspDependencyCheckCvssThresholdValue threshold = visitor.threshold()
+            .map(OWASP_DEPENDENCY_CHECK_FAIL_CVSS_THRESHOLD::value)
+            .orElse(OWASP_DEPENDENCY_CHECK_FAIL_CVSS_THRESHOLD.notSpecifiedValue());
+
+        values.update(usage, threshold);
+      }
     }
   }
 
   /**
-   * Checks if a project uses OWASP Dependency Check Gradle plugin.
+   * Checks if a project uses OWASP Dependency Check Gradle plugin,
+   * and updates a value set with corresponding feature values.
    *
    * @param repository The project's repository.
-   * @return True if the project uses the plugin, false otherwise.
+   * @param values A {@link ValueSet} to be updated.
    */
-  private boolean checkGradle(LocalRepository repository) throws IOException {
-    Optional<InputStream> content = repository.read("build.gradle");
+  private static void checkGradle(LocalRepository repository, ValueSet values) throws IOException {
+    for (Path gradleFile : repository.files(path -> path.getFileName().endsWith(".gradle"))) {
+      Optional<List<String>> something = repository.readLinesOf(gradleFile);
+      if (!something.isPresent()) {
+        return;
+      }
+      List<String> file = something.get();
 
-    if (!content.isPresent()) {
-      return false;
-    }
+      if (foundOwaspDependencyCheckInGradle(file)) {
+        boolean isMainFile = "build.gradle".equals(gradleFile.toString());
 
-    try (BufferedReader reader = new BufferedReader(new InputStreamReader(content.get()))) {
-      String line;
-      while ((line = reader.readLine()) != null) {
-        if (line.trim().contains("org.owasp:dependency-check-gradle")) {
-          return true;
+        OwaspDependencyCheckUsageValue usage
+            = OWASP_DEPENDENCY_CHECK_USAGE.value(isMainFile ? MANDATORY : OPTIONAL);
+
+        OwaspDependencyCheckCvssThresholdValue threshold = lookForThresholdInGradle(file)
+            .map(OWASP_DEPENDENCY_CHECK_FAIL_CVSS_THRESHOLD::value)
+            .orElse(OWASP_DEPENDENCY_CHECK_FAIL_CVSS_THRESHOLD.notSpecifiedValue());
+
+        values.update(usage, threshold);
+
+        if (usage.get() == MANDATORY) {
+          break;
         }
       }
     }
-
-    return false;
   }
 
   /**
-   * Check if a plugin is OWASP Dependency Check plugin.
+   * Searches for a threshold for OWASP Dependency Check in a Gradle file.
    *
-   * @param plugin The plugin to be checked.
-   * @return True if the plugin is OWASP Dependency Check plugin, false otherwise.
+   * @param content The content of the file.
+   * @return The threshold if found.
    */
-  private static boolean isDependencyCheck(Plugin plugin) {
-    return "org.owasp".equals(plugin.getGroupId())
-        && "dependency-check-maven".equals(plugin.getArtifactId());
+  private static Optional<Double> lookForThresholdInGradle(List<String> content) {
+    boolean foundDependencyCheckConfig = false;
+    for (String line : content) {
+      line = line.trim();
+
+      if (line.contains("dependencyCheck")) {
+        foundDependencyCheckConfig = true;
+        continue;
+      }
+
+      if (!foundDependencyCheckConfig) {
+        continue;
+      }
+
+      if ("}".equals(line)) {
+        break;
+      }
+
+      // check for Dependency Check parameters
+      for (Map.Entry<String, String> entry : FAIL_BUILD_CONFIGURATIONS.entrySet()) {
+        String[] parts = line.split("=");
+        if (parts.length != 2) {
+          continue;
+        }
+
+        String name = entry.getKey();
+        if (!parts[0].trim().equals(name)) {
+          continue;
+        }
+
+        String value = parts[1].trim();
+        String type = entry.getValue();
+        Double score = parseScore(value, type);
+
+        // stop once one of the parameters is found
+        return Optional.ofNullable(score);
+      }
+    }
+
+    return Optional.empty();
   }
 
   /**
-   * Check if a report plugin is OWASP Dependency Check plugin.
+   * Checks if a Gradle file calls OWASP Dependency Check.
    *
-   * @param plugin The report plugin to be checked.
-   * @return True if the report plugin is OWASP Dependency Check plugin, false otherwise.
+   * @param content The content of the file.
+   * @return True if the file calls the tool, false otherwise.
    */
-  private static boolean isDependencyCheck(ReportPlugin plugin) {
-    return "org.owasp".equals(plugin.getGroupId())
-        && "dependency-check-maven".equals(plugin.getArtifactId());
+  private static boolean foundOwaspDependencyCheckInGradle(List<String> content) {
+    return content.stream()
+        .filter(line -> line.contains("apply plugin:"))
+        .anyMatch(line -> line.contains("org.owasp.dependencycheck"));
+  }
+
+  /**
+   * Check if groupId and artifactId matches with OWASP Dependency GA.
+   *
+   * @param groupId to be checked.
+   * @param artifactId to be checked.
+   * @return True if the OWASP Dependency GA matches, false otherwise.
+   */
+  private static boolean isDependencyCheck(String groupId, String artifactId) {
+    return "org.owasp".equals(groupId) && "dependency-check-maven".equals(artifactId);
+  }
+
+  /**
+   * Check if a plugin configuration has an attribute that specifies a CVSS threshold,
+   * so that the plugin fails the build if vulnerabilities with higher CVSS score are found.
+   *
+   * @param configuration The plugin configuration to be checked.
+   * @return The CVSS threshold if found, otherwise null.
+   */
+  private static Double cvssScoreFrom(Object configuration) {
+    for (String name : FAIL_BUILD_CONFIGURATIONS.keySet()) {
+      Optional<String> content = parameter(name, configuration);
+
+      if (content.isPresent()) {
+        String type = FAIL_BUILD_CONFIGURATIONS.get(name);
+        return parseScore(content.get(), type);
+      }
+    }
+
+    return null;
+  }
+
+  /**
+   * Parse the value depending on the given type. The type can be 'boolean' or 'number'. 
+   * <ul>
+   *    <li>If the type is 'number', then parse the value.</li>
+   *    <li>If the type is 'boolean' and the value is true, then return 0.0, otherwise null.</li>
+   *    <li>If the type is unknown, an exception is thrown.</li>
+   * </ul>
+   * 
+   * @param value The value to be parsed.
+   * @param type The type of the string value.
+   * @return The parsed value.
+   */
+  private static Double parseScore(String value, String type) {
+    switch (type) {
+      case "number":
+        return Double.parseDouble(value);
+      case "boolean":
+        return Boolean.parseBoolean(value) ? 0.0 : null;
+      default:
+        throw new IllegalArgumentException(String.format("Unknown type: %s", type));
+    }
+  }
+
+  /**
+   * Get the value if a configuration exists.
+   * 
+   * @param name The name of the configuration to find.
+   * @param configuration The list of configurations associated to OWASP Dependency Check plugin.
+   * @return The value of the configuration if found.
+   */
+  private static Optional<String> parameter(String name, Object configuration) {
+    if (configuration instanceof Xpp3Dom == false) {
+      return Optional.empty();
+    }
+
+    Xpp3Dom dom = ((Xpp3Dom) configuration).getChild(name);
+    if (dom == null) {
+      return Optional.empty();
+    }
+
+    return Optional.ofNullable(dom.getValue());
+  }
+
+  /**
+   * Checks whether the plugin will be executed as a mandatory step during the build.
+   * 
+   * @param locations The set of {@link Location}.
+   * @return True if the plugin will be executed as a mandatory step, false otherwise.
+   */
+  private static boolean mandatoryCheckFor(Set<Location> locations) {
+    return locations.size() == 1 && (locations.contains(BUILD) || locations.contains(REPORTING));
   }
 
   /**
@@ -131,22 +325,80 @@ public class UsesOwaspDependencyCheck extends CachedSingleFeatureGitHubDataProvi
   private static class Visitor extends AbstractModelVisitor {
 
     /**
-     * This flag shows whether OWASP Dependency Check tool was found in a POM file or not.
+     * If the plugin is run as a mandatory check.
      */
-    private boolean result = false;
+    private boolean foundMandatory = false;
+    
+    /**
+     * If the plugin is run as an optional check.
+     */
+    private boolean foundOptional = false;
+    
+    /**
+     * The CVSS threshold score to fail the build. 
+     */
+    private Double score; 
 
     @Override
     public void accept(Plugin plugin, Set<Location> locations) {
-      if (isDependencyCheck(plugin)) {
-        result = true;
-      }
+      accept(plugin.getGroupId(), plugin.getArtifactId(), plugin.getConfiguration(), locations);
     }
 
     @Override
     public void accept(ReportPlugin plugin, Set<Location> locations) {
-      if (isDependencyCheck(plugin)) {
-        result = true;
+      accept(plugin.getGroupId(), plugin.getArtifactId(), plugin.getConfiguration(), locations);
+    }
+
+    /**
+     * Check if a plugin is OWASP Dependency Check plugin.
+     *
+     * @param groupId A group id of the plugins.
+     * @param artifactId An artifact id of the plugin.
+     * @param configuration A configuration of the plugin.
+     * @param locations Locations of the plugin.
+     */
+    private void accept(
+        String groupId, String artifactId, Object configuration, Set<Location> locations) {
+
+      if (foundMandatory) {
+        return;
       }
+
+      if (isDependencyCheck(groupId, artifactId)) {
+        if (mandatoryCheckFor(locations)) {
+          foundMandatory = true;
+        } else {
+          foundOptional = true;
+        }
+
+        score = cvssScoreFrom(configuration);
+      }
+    }
+
+    /**
+     * Get {@link OwaspDependencyCheckUsage} based on where the plugin is found.
+     *
+     * @return {@link OwaspDependencyCheckUsage} based on where the plugin is found.
+     */
+    OwaspDependencyCheckUsage usage() {
+      if (foundMandatory) {
+        return MANDATORY;
+      }
+
+      if (foundOptional) {
+        return OPTIONAL;
+      }
+
+      return NOT_USED;
+    }
+
+    /**
+     * Get a CVSS threshold if found.
+     *
+     * @return The threshold.
+     */
+    Optional<Double> threshold() {
+      return Optional.ofNullable(score);
     }
   }
 }
