+++ src/test/java/com/sap/sgs/phosphor/fosstars/data/github/UsesOwaspDependencyCheckTest.java
@@ -1,30 +1,49 @@
 package com.sap.sgs.phosphor.fosstars.data.github;
 
 import static com.sap.sgs.phosphor.fosstars.data.github.TestGitHubDataFetcherHolder.TestGitHubDataFetcher.addForTesting;
-import static com.sap.sgs.phosphor.fosstars.data.github.UsesOwaspDependencyCheck.USES_OWASP_DEPENDENCY_CHECK;
+import static com.sap.sgs.phosphor.fosstars.model.feature.oss.OssFeatures.OWASP_DEPENDENCY_CHECK_FAIL_CVSS_THRESHOLD;
+import static com.sap.sgs.phosphor.fosstars.model.feature.oss.OssFeatures.OWASP_DEPENDENCY_CHECK_USAGE;
+import static com.sap.sgs.phosphor.fosstars.model.value.OwaspDependencyCheckUsage.MANDATORY;
+import static com.sap.sgs.phosphor.fosstars.model.value.OwaspDependencyCheckUsage.NOT_USED;
+import static com.sap.sgs.phosphor.fosstars.model.value.OwaspDependencyCheckUsage.OPTIONAL;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+import com.sap.sgs.phosphor.fosstars.model.Feature;
 import com.sap.sgs.phosphor.fosstars.model.Value;
 import com.sap.sgs.phosphor.fosstars.model.ValueSet;
+import com.sap.sgs.phosphor.fosstars.model.value.CVSS;
+import com.sap.sgs.phosphor.fosstars.model.value.OwaspDependencyCheckCvssThresholdValue;
+import com.sap.sgs.phosphor.fosstars.model.value.OwaspDependencyCheckUsage;
 import com.sap.sgs.phosphor.fosstars.model.value.ValueHashSet;
 import com.sap.sgs.phosphor.fosstars.tool.github.GitHubProject;
 import com.sap.sgs.phosphor.fosstars.tool.github.GitHubProjectValueCache;
 import java.io.IOException;
 import java.io.InputStream;
+import java.nio.file.Paths;
+import java.util.Collections;
+import java.util.List;
 import java.util.Optional;
+import java.util.Set;
+import org.apache.commons.io.IOUtils;
 import org.junit.Test;
 
 public class UsesOwaspDependencyCheckTest extends TestGitHubDataFetcherHolder {
+
+  private static final Double NOT_SPECIFIED = null;
 
   @Test
   public void testMavenWithOwaspDependencyCheckInBuild() throws IOException {
     try (InputStream is = getClass()
         .getResourceAsStream("MavenWithOwaspDependencyCheckInBuild.xml")) {
 
-      checkValue(createProvider(is, "pom.xml"), true);
+      ValueSet values = values(createProvider(is, "pom.xml"));
+      checkUsage(MANDATORY, values);
+      checkThreshold(NOT_SPECIFIED, values);
     }
   }
 
@@ -33,34 +52,42 @@ public class UsesOwaspDependencyCheckTest extends TestGitHubDataFetcherHolder {
     try (InputStream is = getClass()
         .getResourceAsStream("MavenWithOwaspDependencyCheckInBuildPluginManagement.xml")) {
 
-      checkValue(createProvider(is, "pom.xml"), true);
+      ValueSet values = values(createProvider(is, "pom.xml"));
+      checkUsage(OPTIONAL, values);
+      checkThreshold(NOT_SPECIFIED, values);
     }
   }
 
   @Test
   public void testMavenWithOwaspDependencyCheckInProfilesBuild() throws IOException {
-    try (InputStream is = getClass()
-        .getResourceAsStream("MavenWithOwaspDependencyCheckInProfilesBuild.xml")) {
+    try (InputStream is =
+        getClass().getResourceAsStream("MavenWithOwaspDependencyCheckInProfilesBuild.xml")) {
 
-      checkValue(createProvider(is, "pom.xml"), true);
+      ValueSet values = values(createProvider(is, "pom.xml"));
+      checkUsage(OPTIONAL, values);
+      checkThreshold(7.0, values);
     }
   }
 
   @Test
   public void testMavenWithOwaspDependencyCheckInProfilesReporting() throws IOException {
-    try (InputStream is = getClass()
-        .getResourceAsStream("MavenWithOwaspDependencyCheckInProfilesReporting.xml")) {
+    try (InputStream is =
+        getClass().getResourceAsStream("MavenWithOwaspDependencyCheckInProfilesReporting.xml")) {
 
-      checkValue(createProvider(is, "pom.xml"), true);
+      ValueSet values = values(createProvider(is, "pom.xml"));
+      checkUsage(OPTIONAL, values);
+      checkThreshold(0.0, values);
     }
   }
 
   @Test
   public void testMavenWithOwaspDependencyCheckInReporting() throws IOException {
-    try (InputStream is = getClass()
-        .getResourceAsStream("MavenWithOwaspDependencyCheckInProfilesReporting.xml")) {
+    try (InputStream is =
+        getClass().getResourceAsStream("MavenWithOwaspDependencyCheckInReporting.xml")) {
 
-      checkValue(createProvider(is, "pom.xml"), true);
+      ValueSet values = values(createProvider(is, "pom.xml"));
+      checkUsage(MANDATORY, values);
+      checkThreshold(1.3, values);
     }
   }
 
@@ -69,16 +96,42 @@ public class UsesOwaspDependencyCheckTest extends TestGitHubDataFetcherHolder {
     try (InputStream is = getClass()
         .getResourceAsStream("MavenWithoutOwaspDependencyCheck.xml")) {
 
-      checkValue(createProvider(is, "pom.xml"), false);
+      ValueSet values = values(createProvider(is, "pom.xml"));
+      checkUsage(NOT_USED, values);
+      checkThreshold(NOT_SPECIFIED, values);
+    }
+  }
+  
+  @Test
+  public void testMavenWitOwaspDependencyCheckBuildAndProfile() throws IOException {
+    try (InputStream is = getClass()
+        .getResourceAsStream("MavenWithOwaspDependencyCheckInBuildAndProfile.xml")) {
+
+      ValueSet values = values(createProvider(is, "pom.xml"));
+      checkUsage(MANDATORY, values);
+      checkThreshold(6.0, values);
     }
   }
 
   @Test
-  public void testGradleWithOwaspDependencyCheck() throws IOException {
+  public void testGradleWithMandatoryOwaspDependencyCheck() throws IOException {
     try (InputStream is = getClass()
         .getResourceAsStream("GradleWithOwaspDependencyCheck.gradle")) {
 
-      checkValue(createProvider(is, "build.gradle"), true);
+      ValueSet values = values(createProvider(is, "build.gradle"));
+      checkUsage(MANDATORY, values);
+      checkThreshold(NOT_SPECIFIED, values);
+    }
+  }
+
+  @Test
+  public void testGradleWithOptionalOwaspDependencyCheck() throws IOException {
+    try (InputStream is = getClass()
+        .getResourceAsStream("GradleWithOwaspDependencyCheck.gradle")) {
+
+      ValueSet values = values(createProvider(is, "other/build.gradle"));
+      checkUsage(OPTIONAL, values);
+      checkThreshold(NOT_SPECIFIED, values);
     }
   }
 
@@ -87,7 +140,46 @@ public class UsesOwaspDependencyCheckTest extends TestGitHubDataFetcherHolder {
     try (InputStream is = getClass()
         .getResourceAsStream("GradleWithoutOwaspDependencyCheck.gradle")) {
 
-      checkValue(createProvider(is, "build.gradle"), false);
+      ValueSet values = values(createProvider(is, "build.gradle"));
+      checkUsage(NOT_USED, values);
+      checkThreshold(NOT_SPECIFIED, values);
+    }
+  }
+
+  @Test
+  public void testGradleWithMandatoryOwaspDependencyCheckWithFailBuildOnCvss() throws IOException {
+    try (InputStream is = getClass()
+        .getResourceAsStream("GradleWithOwaspDependencyCheckWithFailBuildOnCvss.gradle")) {
+
+      ValueSet values = values(createProvider(is, "build.gradle"));
+      checkUsage(MANDATORY, values);
+      checkThreshold(5.3, values);
+    }
+  }
+
+  @Test
+  public void testGradleWithMandatoryOwaspDependencyCheckWithFailBuildOnAnyIssueTrue()
+      throws IOException {
+
+    try (InputStream is = getClass()
+        .getResourceAsStream("GradleWithOwaspDependencyCheckWithFailBuildOnAnyIssueTrue.gradle")) {
+
+      ValueSet values = values(createProvider(is, "build.gradle"));
+      checkUsage(MANDATORY, values);
+      checkThreshold(CVSS.MIN, values);
+    }
+  }
+
+  @Test
+  public void testGradleWithMandatoryOwaspDependencyCheckWithBuildOnAnyIssueFalse()
+      throws IOException {
+
+    try (InputStream is = getClass()
+        .getResourceAsStream("GradleWithOwaspDependencyCheckWithFailBuildOnAnyIssueFalse.gradle")) {
+
+      ValueSet values = values(createProvider(is, "build.gradle"));
+      checkUsage(MANDATORY, values);
+      checkThreshold(NOT_SPECIFIED, values);
     }
   }
 
@@ -95,7 +187,15 @@ public class UsesOwaspDependencyCheckTest extends TestGitHubDataFetcherHolder {
       throws IOException {
 
     final LocalRepository repository = mock(LocalRepository.class);
-    when(repository.read(filename)).thenReturn(Optional.of(is));
+
+    List<String> content = IOUtils.readLines(is);
+
+    when(repository.read(filename))
+        .thenReturn(Optional.of(IOUtils.toInputStream(String.join("\n", content))));
+    when(repository.readLinesOf(Paths.get(filename)))
+        .thenReturn(Optional.of(content));
+    when(repository.files(any()))
+        .thenReturn(Collections.singletonList(Paths.get(filename)));
 
     GitHubProject project = new GitHubProject("org", "test");
     addForTesting(project, repository);
@@ -106,20 +206,37 @@ public class UsesOwaspDependencyCheckTest extends TestGitHubDataFetcherHolder {
     return provider;
   }
 
-  private static void checkValue(UsesOwaspDependencyCheck provider, boolean expectedValue)
-      throws IOException {
-
+  private static ValueSet values(UsesOwaspDependencyCheck provider) throws IOException {
     GitHubProject project = new GitHubProject("org", "test");
 
     ValueSet values = new ValueHashSet();
     provider.update(project, values);
 
-    assertEquals(1, values.size());
-    assertTrue(values.has(USES_OWASP_DEPENDENCY_CHECK));
-    assertTrue(values.of(USES_OWASP_DEPENDENCY_CHECK).isPresent());
-
-    Value<Boolean> value = values.of(USES_OWASP_DEPENDENCY_CHECK).get();
-    assertEquals(expectedValue, value.get());
+    Set<Feature> features = provider.supportedFeatures();
+    assertEquals(features.size(), values.size());
+    assertTrue(values.containsAll(features));
+    return values;
   }
 
+  private static void checkUsage(OwaspDependencyCheckUsage expected, ValueSet values) {
+    Optional<Value> value = values.of(OWASP_DEPENDENCY_CHECK_USAGE);
+    assertTrue(value.isPresent());
+    Value usage = value.get();
+    assertEquals(expected, usage.get());
+  }
+
+  private static void checkThreshold(Double expected, ValueSet values) {
+    Optional<Value> value = values.of(OWASP_DEPENDENCY_CHECK_FAIL_CVSS_THRESHOLD);
+    assertTrue(value.isPresent());
+    Value n = value.get();
+    assertTrue(n instanceof OwaspDependencyCheckCvssThresholdValue);
+    OwaspDependencyCheckCvssThresholdValue threshold = (OwaspDependencyCheckCvssThresholdValue) n;
+
+    if (expected == null) {
+      assertFalse(threshold.specified());
+    } else {
+      assertTrue(threshold.specified());
+      assertEquals(expected, threshold.get());
+    }
+  }
 }
